+++ src/main/java/com/gromed/demo/repository/CompteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompteRepository extends JpaRepository<Compte, Integer> {
+public interface CompteRepository extends JpaRepository<Compte, Long> {
 }
